# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Starter API",
    "contact": {}
  },
  "host": "localhost:8080.",
  "paths": {
...
    "/api/v1/user": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Создание пользователя",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createUserRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createUserResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список пользователей",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getUsersResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/users/:id": {
      "get": {
        "produces": ["application/json"],
        "summary": "Возвращает информацию о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id пользователя",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getUserResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Удаление информации о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id пользователя",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "patch": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Обновление информации о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id пользователя",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.patchUserRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "model.ResponseError": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/model.ResponseErrorField"
        },
        "success": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "model.ResponseErrorField": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "details": {
          "type": "string"
        }
      }
    },
    "model.ResponseSuccess": {
      "type": "object",
      "properties": {
        "data": {},
        "success": {
          "type": "boolean",
          "default": true
        }
      }
    },
...
    "requests.createUserRequestData": {
      "type": "object",
      "required": ["password", "phone", "role"],
      "properties": {
        "email": {
          "type": "string"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
...
    "requests.patchUserRequestData": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
...
    "responses.createUserResponseData": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/responses.createUserResponseUser"
        }
      }
    },
    "responses.createUserResponseUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
...
    "responses.getUserResponseData": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/responses.getUserResponseUser"
        }
      }
    },
    "responses.getUserResponseUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
    "responses.getUsersResponseData": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getUsersResponseUser"
          }
        }
      }
    },
    "responses.getUsersResponseUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    }
  }
}